requirements.txt
# code_snippets/requirements.txt

numpy
scikit-learn
tensorflow
keras
torch
PyYAML
joblib
